simulation view skeleton
