+++ client/styles/globals.css
@@ -219,3 +219,7 @@ body {
     flex-grow: 1;
   }
 }
+
+.danger {
+  background-color: #e74c3c!important;
+}
